refactor: split grid state file
